Add files via upload
This is a generic heap that I created for class. I also created three comparators. The first is a string comparator that compares strings given their size and in reverse lexicographical order. It prints the smallest words first and then in reverse alphabetical order if they are the same size. The second compares integers to see if they are even or odd, and their sizes. It prints even numbers first, smallest to large, then odd numbers smallest to largest. The third compares integers again and prints odd numbers first in priority, smallest to largest, and then even numbers, with priority being largest to smallest.